+++ lib/drivers/io.c
@@ -25,12 +25,12 @@ void set_bit_mask(volatile uint32_t* bits, uint32_t mask, uint32_t value)
     *bits = org | (value & mask);
 }
 
-uint32_t get_bit_idx(volatile uint32_t* bits, size_t idx)
+uint32_t get_bit_idx(volatile uint32_t* bits, uint32_t idx)
 {
     return ((*bits) & (1 << idx)) >> idx;
 }
 
-void set_bit_idx(volatile uint32_t* bits, size_t idx, uint32_t value)
+void set_bit_idx(volatile uint32_t* bits, uint32_t idx, uint32_t value)
 {
     uint32_t org = (*bits) & ~(1 << idx);
     *bits = org | (value << idx);
